+++ src/Common/Service/SettingsService.cs
@@ -1,6 +1,9 @@
 ﻿using Common.Database;
 using Common.Observe;
+using Common.Stateful;
+using Microsoft.Extensions.Caching.Memory;
 using Serilog;
+using System;
 using System.Threading.Tasks;
 
 namespace Common.Service;
@@ -8,30 +11,70 @@ namespace Common.Service;
 public interface ISettingsService
 {
 	Task<Settings> GetSettingsAsync();
-	Task UpdateSettings(Settings settings);
+	Task UpdateSettingsAsync(Settings settings);
+	PelotonApiAuthentication GetPelotonApiAuthentication(string pelotonEmail);
+	void SetPelotonApiAuthentication(PelotonApiAuthentication authentication);
+	void ClearPelotonApiAuthentication(string pelotonEmail);
 }
 
 public class SettingsService : ISettingsService
 {
 	private static readonly ILogger _logger = LogContext.ForClass<SettingsService>();
+	private static readonly object _lock = new object();
+	private static readonly string PelotonApiAuthKey = "PelotonApiAuth";
 
 	private readonly ISettingsDb _db;
+	private readonly IMemoryCache _cache;
 
-	public SettingsService(ISettingsDb db)
+	public SettingsService(ISettingsDb db, IMemoryCache cache)
 	{
 		_db = db;
+		_cache = cache;
 	}
 
-	public Task<Settings> GetSettingsAsync()
+	public PelotonApiAuthentication GetPelotonApiAuthentication(string pelotonEmail)
+	{
+		using var tracing = Tracing.Trace($"{nameof(SettingsService)}.{nameof(GetPelotonApiAuthentication)}");
+
+		lock (_lock)
+		{
+			var key = $"{PelotonApiAuthKey}:{pelotonEmail}";
+			return _cache.Get<PelotonApiAuthentication>(key);
+		}
+	}
+
+	public async Task<Settings> GetSettingsAsync()
 	{
 		using var tracing = Tracing.Trace($"{nameof(SettingsService)}.{nameof(GetSettingsAsync)}");
 
-		return _db.GetSettingsAsync();
+		return (await _db.GetSettingsAsync()) ?? new Settings();
 	}
 
-	public async Task UpdateSettings(Settings updatedSettings)
+	public void SetPelotonApiAuthentication(PelotonApiAuthentication authentication)
 	{
-		using var tracing = Tracing.Trace($"{nameof(SettingsService)}.{nameof(UpdateSettings)}");
+		using var tracing = Tracing.Trace($"{nameof(SettingsService)}.{nameof(SetPelotonApiAuthentication)}");
+
+		lock (_lock)
+		{
+			var key = $"{PelotonApiAuthKey}:{authentication.Email}";
+			_cache.Set(key, authentication, new MemoryCacheEntryOptions() { AbsoluteExpirationRelativeToNow = TimeSpan.FromMinutes(15) });
+		}
+	}
+
+	public void ClearPelotonApiAuthentication(string pelotonEmail)
+	{
+		using var tracing = Tracing.Trace($"{nameof(SettingsService)}.{nameof(ClearPelotonApiAuthentication)}");
+
+		lock (_lock)
+		{
+			var key = $"{PelotonApiAuthKey}:{pelotonEmail}";
+			_cache.Remove(key);
+		}
+	}
+
+	public async Task UpdateSettingsAsync(Settings updatedSettings)
+	{
+		using var tracing = Tracing.Trace($"{nameof(SettingsService)}.{nameof(UpdateSettingsAsync)}");
 
 		var originalSettings = await _db.GetSettingsAsync();
 
@@ -40,7 +83,9 @@ public class SettingsService : ISettingsService
 
 		if (updatedSettings.Peloton.Password is null)
 			updatedSettings.Peloton.Password = originalSettings.Peloton.Password;
-
+		
+		ClearPelotonApiAuthentication(originalSettings.Peloton.Email);
+		ClearPelotonApiAuthentication(updatedSettings.Peloton.Email);
 		await _db.UpsertSettingsAsync(updatedSettings);
 	}
 }