+++ src/Common/Stateful/PelotonApiAuthentication.cs
@@ -1,0 +1,10 @@
+﻿namespace Common.Stateful
+{
+	public class PelotonApiAuthentication
+	{
+		public string Email { get; set; }
+		public string Password { get; set; }
+		public string UserId { get; set; }
+		public string SessionId { get; set; }
+	}
+}