+++ src/Conversion/ConvertStatus.cs
@@ -2,7 +2,14 @@
 {
 	public class ConvertStatus
 	{
-		public bool Success { get; set; }
+		public ConversionResult Result { get; set; }
 		public string ErrorMessage { get; set; }
+	}
+
+	public enum ConversionResult
+	{
+		Success = 0,
+		Skipped = 10,
+		Failed = 20,	
 	}
 }