+++ src/Conversion/TcxConverter.cs
@@ -17,7 +17,7 @@ namespace Conversion
 
 		public override ConvertStatus Convert(P2GWorkout workout)
 		{
-			if (!_config.Format.Tcx) return new ConvertStatus() { Success = true, ErrorMessage = "Tcx format disabled in config." };
+			if (!_config.Format.Tcx) return new ConvertStatus() {  Result = ConversionResult.Skipped };
 
 			return base.ConvertForFormat(FileFormat.Fit, workout);
 		}