+++ src/WebUI/Domain/ErrorResponse.cs
@@ -1,4 +1,0 @@
-﻿public class ErrorResponse
-{
-	public string Message { get; set; }
-}