+++ src/WebUI/Domain/PelotonWorkoutsRequest.cs
@@ -1,8 +1,0 @@
-﻿using Common.Dto.Peloton;
-
-namespace WebUI.Domain;
-
-public class RecentWorkoutsGetResponse
-{
-	public ICollection<RecentWorkout> Items { get; set; }
-}