+++ ReleaseNotes.md
@@ -1,7 +1,13 @@
-## Changes
+## Features
 
-- [#316] Bump dependencies
+- [#289] Add paging support to Sync UI page
+	- API and WebUI must be updated together to atleast 3.1.0
 
 ## Fixes
 
-- [#331] Fixed - System.IO.IOException: The process cannot access the file 'deviceInfo.xml' because it is being used by another process. 
+- [#331] Fixed - System.IO.IOException: The process cannot access the file 'deviceInfo.xml' because it is being used by another process.
+- [#289] Fixed - Better logs and messaging when no convert Formats are configured in Settings
+
+## Changes
+
+- [#316] Bump dependencies